Backend/user_domain/authService/validateToken/requirements.txt
flask
requests
redis
python-dotenv
python-jose[cryptography]
werkzeug
flask-cors
